backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
spotipy==2.23.0
bcrypt==4.0.1
python-dateutil==2.8.2
pandas==2.2.0
numpy==1.26.4
google-generativeai==0.3.2
google-generativeai==0.3.2
groq==0.31.0
